refactor(codecov): Move pageInfo serializer to be shared
